[week4] 일단 구현 완료

+++ asset/todolist-jq.js
@@ -1,30 +1,91 @@
 var TODOSync = {
     URL : "http://128.199.76.9:8002/nickname",
+    init : function () {
+        window.addEventListener("online", this.onoffLineListener);
+        window.addEventListener("offline", this.onoffLineListener);
+    },
+    onoffLineListener : function () {
+        console.log("hello");
+        document.getElementById('header').classList[navigator.onLine?"remove":"add"]("offline");
+        if(navigator.onLine){
+            //서버로 Sync 맞추기
+        }
+    },
     get : function(callback) {
-        $.ajax({method: "GET", url: this.URL}).done(callback);
+        if (navigator.onLine) {
+            $.ajax({method: "GET", url: this.URL}).done(callback);
+        }else{
+            //local db
+        }
     },
     add : function(todo,callback) {
-        $.ajax({method: "PUT", url: this.URL, data: "todo="+todo}).done(callback);
+        if (navigator.onLine) {
+            $.ajax({method: "PUT", url: this.URL, data: "todo="+todo}).done(callback);
+        };
     },
     completed : function(param,callback) {
-        $.ajax({method: "POST", url: this.URL+'/'+param.key, data:"completed="+param.completed}).done(callback);
+        if (navigator.onLine) {
+            $.ajax({method: "POST", url: this.URL+'/'+param.key, data:"completed="+param.completed}).done(callback);
+        };
     },
     remove : function(param,callback) {
-        $.ajax({method: "DELETE", url: this.URL+'/'+param.key}).done(callback);
+        if (navigator.onLine) {
+            $.ajax({method: "DELETE", url: this.URL+'/'+param.key}).done(callback);
+        };
     }
 }
 
 var TODO = {
-    oInput: null,
+    oInput: $("#new-todo"),
+    oList : $("#todo-list"),
     ENTER_KEYCODE : 13,
+    ShowState : {index:0, className:""},
     init: function () {
-        this.oInput = $("#new-todo");
-        // 이건 왜 않먹히는 걸까요?
-        // this.oInput = $("#new-todo").get(0);
         this.oInput.on("keydown", this.add.bind(this));
         $("#todo-list").on("click", ".toggle", this.completed);
         $("#todo-list").on("click", ".destroy", this.remove);
         this.get();
+        $("#filters").on("click", "a", this.changeStateFilter);
+        window.addEventListener("popState", this.changeUrlFilter);
+    },
+    changeUrlFilter : function () {
+        
+    }
+    changeStateFilter : function (e) {
+        var href = this.getAttribute("href");
+        // console.log(href)
+        e.preventDefault();
+        if(href === "index.html"){
+            TODO.selectView({
+                index:0, className:"", 
+                method: "all", url: href});
+        }else if(href === "active"){
+            TODO.selectView({
+                index:1, className:"all-active", 
+                method:"active", url: href});
+        }else if(href === "completed"){
+            TODO.selectView({
+                index:2, className:"all-completed", 
+                method:"completed", url: href});
+        }
+    },
+    selectView : function (nowState) {
+        var navigatorList = $("#filters a");
+        $(navigatorList.get(this.ShowState.index)).removeClass("selected");
+        $("#todo-list").removeClass(this.ShowState.className);
+
+        $(navigatorList.get(nowState.index)).addClass("selected");
+        $("#todo-list").addClass(nowState.className);
+        this.ShowState = nowState;
+        // 히스토리 추가
+        // 뒤로가기 이벤트 변경
+        console.log(nowState);
+        history.pushState(nowState.method, null, nowState.href);
+    },
+    selectedNavigator : function (index) {
+        var navigatorList = $("#filters a");
+        $(navigatorList.get(this.selectedIndex)).removeClass("selected");
+        $(navigatorList.get(index)).addClass("selected");
     },
     build: function(context) {
     	var html = $("#entry-template").html();
@@ -32,7 +93,7 @@ var TODO = {
     	this.build = template;
         return template(context);
     },
-    get : function () {
+    get : function () { 
         TODOSync.get(function (json) {
             $("#todo-list").append(json.map(function (obj) {
                 return TODO.build({target: obj.todo, key: obj.id, completed: obj.completed});
@@ -70,7 +131,6 @@ var TODO = {
     },
     add : function (e) {
         if (e.keyCode !== this.ENTER_KEYCODE) {return}
-
         var todo = TODO.oInput.val();
         TODOSync.add(todo, function (json) {
             $("#todo-list").prepend(TODO.build({target: todo, key: json.insertId}));
@@ -87,5 +147,6 @@ var TODO = {
 
 $(function () {
     TODO.init();
+    TODOSync.init();
 });
